reflect review

+++ src/common.c
@@ -994,7 +994,7 @@ void stlink_print_data(stlink_t * sl) {
     if (sl->verbose > 2)
         fprintf(stdout, "data_len = %d 0x%x\n", sl->q_len, sl->q_len);
 
-    int i;
+    size_t i;
     for (i = 0; i < sl->q_len; i++) {
         if (i % 16 == 0) {
             /*
@@ -1311,7 +1311,7 @@ static bool stlink_fread_ihex_writeline(struct stlink_fread_ihex_worker_arg* the
     if(9 != fprintf(the_arg->file, ":%02X%04X00", count, (addr & 0x0000FFFF)))
         return false;
 
-    uint8_t i;
+    size_t i;
     for(i = 0; i < count; ++i) {
         uint8_t b = the_arg->buf[i];
         sum += b;
@@ -1340,7 +1340,7 @@ static bool stlink_fread_ihex_init(struct stlink_fread_ihex_worker_arg* the_arg,
 static bool stlink_fread_ihex_worker(void* arg, uint8_t* block, ssize_t len) {
     struct stlink_fread_ihex_worker_arg* the_arg = (struct stlink_fread_ihex_worker_arg*)arg;
 
-    ssize_t i;
+    size_t i;
     for(i = 0; i < len; ++i) {
         if(the_arg->buf_pos == sizeof(the_arg->buf)) { // line is full
             if(!stlink_fread_ihex_writeline(the_arg)) return false;
@@ -2043,7 +2043,7 @@ int stlink_write_flash(stlink_t *sl, stm32_addr_t addr, uint8_t* base, uint32_t 
 // note: length not checked
 static uint8_t stlink_parse_hex(const char* hex) {
     uint8_t d[2];
-    int i;
+    size_t i;
     for(i = 0; i < 2; ++i) {
         char c = *(hex + i);
         if(c >= '0' && c <= '9') d[i] = c - '0';
@@ -2061,7 +2061,7 @@ int stlink_parse_ihex(const char* path, uint8_t erased_pattern, uint8_t * * mem,
     uint32_t end = 0;
     bool eof_found = false;
 
-    int scan;
+    size_t scan;
     for(scan = 0; (res == 0) && (scan < 2); ++scan) { // parse file two times - first to find memory range, second - to fill it
         if(scan == 1) {
             if(!eof_found) {
@@ -2144,7 +2144,7 @@ int stlink_parse_ihex(const char* path, uint8_t erased_pattern, uint8_t * * mem,
                         if(e > end) end = e;
                     }
                     else {
-			uint8_t i;
+			size_t i;
                         for(i = 0; i < reclen; ++i) {
                             uint8_t b = stlink_parse_hex(line + 9 + i*2);
                             uint32_t addr = lba + offset + i;
