+++ src/sg.c
@@ -151,7 +151,7 @@ static int dump_CDB_command(uint8_t *cdb, uint8_t cdb_len) {
     char dbugblah[100];
     char *dbugp = dbugblah;
     dbugp += sprintf(dbugp, "Sending CDB [");
-    uint8_t i;
+    size_t i;
     for (i = 0; i < cdb_len; i++) {
         dbugp += sprintf(dbugp, " %#02x", (unsigned int) cdb[i]);
     }
@@ -602,7 +602,7 @@ int _stlink_sg_read_all_regs(stlink_t *sl, struct stlink_reg *regp) {
 
     // 0-3 | 4-7 | ... | 60-63 | 64-67 | 68-71   | 72-75      | 76-79 | 80-83
     // r0  | r1  | ... | r15   | xpsr  | main_sp | process_sp | rw    | rw2
-    int i;
+    size_t i;
     for (i = 0; i < 16; i++) {
         regp->r[i] = read_uint32(sl->q_buf, 4 * i);
         if (sl->verbose > 1)
